LWPx::ParanoidAgent compatibility class
t/compat/ is imported from LWPx-ParanoidAgent and only slightly modified
to accommodate our testing infrastructure.  30-mirror.t was dropped
because of dubious usefulness/correctness.  I suspect it hasn't been run
against LWPx-ParanoidAgent in quite a while since there are a number of
manual hoops to jump through to do so.  Even if it had, it doesn't seem
to produce the correct result anymore (is_success is not true, as the
response is a "500 read timeout").

t/compat/ is additionally marked for release testing only.

All tests pass for me when running:

    env RELEASE_TESTING=1 prove -wlvr t/